refactor(tiling): address #207 review findings
Split TilingEngine+Stash.swift past the 350-line ceiling along the
park-vs-restore seam (restore/capture family moves to
TilingEngine+StashRestore.swift). Document the three deliberate
non-clients of spaceSwitchRetile() at the authority and at the
settle site (instant park + in-flight skip beats a late re-slide).
Reword the rejected sequenced variant's 'phase 1' vocabulary, fence
cancelAll's no-snap stranding hazard, and correct the lua-reference
perf sentence (frames apply off-main; slow apps fall behind rather
than block).
